+++ SDL-Game/src/Vector2.h
@@ -1,30 +1,31 @@
 #pragma once
 
+#include "DataTypes.h"
 #include <stdexcept>
 
 struct Vector2
 {
-	float x, y;
+	FLOAT32 x, y;
 
 	Vector2();
-	explicit Vector2(float x1, float y1);
+	explicit Vector2(FLOAT32 x1, FLOAT32 y1);
 	Vector2(const Vector2& other);
 	Vector2(Vector2&& other) noexcept;
 
-	float SqrMagnitude() const;
-	float Magnitude() const;
+	FLOAT32 SqrMagnitude() const;
+	FLOAT32 Magnitude() const;
 
 	Vector2 Normalised() const;
 	Vector2& Normalise();
 
-	Vector2 Rotated(float degrees) const;
-	Vector2& Rotate(float degrees);
+	Vector2 Rotated(FLOAT32 degrees) const;
+	Vector2& Rotate(FLOAT32 degrees);
 
-	static float Dot(const Vector2& a, const Vector2& b);
-	static float Cross(const Vector2& a, const Vector2& b);
+	static FLOAT32 Dot(const Vector2& a, const Vector2& b);
+	static FLOAT32 Cross(const Vector2& a, const Vector2& b);
 
-	float Get(int index) const;
-	float& Get(int index);
+	FLOAT32 Get(SIZE_T index) const;
+	FLOAT32& Get(SIZE_T index);
 
 	Vector2 operator+(const Vector2& other) const { return Vector2(x + other.x, y + other.y); }
 	Vector2 operator-(const Vector2& other) const { return Vector2(x - other.x, y - other.y); }
@@ -32,13 +33,13 @@ struct Vector2
 	Vector2 operator*(float scale) const { return Vector2(x * scale, y * scale); }
 	Vector2 operator/(float scale) const { return Vector2(x / scale, y / scale); }
 
-	float operator[](int index) const
+	FLOAT32 operator[](SIZE_T index) const
 	{
 		if (index == 0) return x;
 		else if (index == 1) return y;
 		else throw new std::out_of_range("Index out of range. Index needs to be 0 for x or 1 for y");
 	}
-	float& operator[](int index)
+	float& operator[](SIZE_T index)
 	{
 		if (index == 0) return x;
 		else if (index == 1) return y;
@@ -58,13 +59,13 @@ struct Vector2
 		return *this;
 	}
 
-	Vector2& operator*=(float scale)
+	Vector2& operator*=(FLOAT32 scale)
 	{
 		x *= scale;
 		y *= scale;
 		return *this;
 	}
-	Vector2& operator/=(float scale)
+	Vector2& operator/=(FLOAT32 scale)
 	{
 		x /= scale;
 		y /= scale;
